split ownership
