+++ packages/visualization/src/components/ScatterPlot/ScatterPlot.stories.tsx
@@ -25,7 +25,6 @@ type Point = {
     y: number;
     color: string;
     shape?: "circle" | "triangle";
-    label?: string ;
 };
 
 export default meta;
@@ -35,7 +34,8 @@ type Story = StoryObj<typeof meta>;
 const scatterData: Point[] = [
     { x: 1, y: 2, color: 'red' },
     { x: 3, y: 4, color: 'blue' },
-    { x: 5, y: 6, color: 'green'},
+    { x: 5, y: 6, color: 'green' },
+    { x:-5, y: -6, color: 'green' },
 ];
 
 
@@ -131,6 +131,17 @@ export const CustomTooltip: Story = {
         pointData: scatterData,
         loading: false,
         miniMap: miniMap,
+        backgroundGradient:{
+    colorScale: ["red", "white", "blue"],
+    legend: {
+      label: "L2FC TRvUT",
+      minLabel: "-0.8",
+      midLabel: "0",
+      maxLabel: "0.8",
+    }
+  },
+  originLine: true,
+  border: true,
         leftAxisLabel: "Y-Axis Label",
         bottomAxisLabel: "X-Axis Label",
         tooltipBody: (point) => (
